backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-jwt-extended==4.7.1
flask-cors==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
gunicorn==23.0.0
